+++ dispatch_tables_fallback_test.go
@@ -12,6 +12,8 @@ var allFlatTables = map[string][]any{
 	"Bytes.RunStartsI32":             {tblBytesRunStartsI32[0]},
 	"Bytes.RunStartsI64":             {tblBytesRunStartsI64[0]},
 	"Bytes.RunStartsU8":              {tblBytesRunStartsU8[0]},
+	"Bytes.Adler32":                  {tblBytesAdler32[0]},
+	"Bytes.CRC32C":                   {tblBytesCRC32C[0]},
 	"Convert.QMatMulI8":              {tblConvertQMatMulI8[0]},
 	"Convert.RequantizeI8":           {tblConvertRequantizeI8[0]},
 	"Mask.All":                       {tblMaskAll[0]},